Core/requirements.txt
# --- Core ML stack ---
torch==2.4.1
transformers==4.45.2
pdfplumber==0.11.4
Pillow==10.4.0
nltk==3.9.1
regex==2024.11.6
tqdm==4.66.5
numpy==1.26.4
pandas==2.2.3
faiss-cpu==1.8.0